feat: extract GameModePanel to shared component with vertical layout
